+++ Cantera/src/equil/vcs_SpeciesProperties.cpp
@@ -19,8 +19,8 @@ namespace VCSnonideal {
  *
  * constructor():
  */
-vcs_SpeciesProperties::vcs_SpeciesProperties(int indexPhase, 
-					     int indexSpeciesPhase,
+vcs_SpeciesProperties::vcs_SpeciesProperties(size_t indexPhase,
+					     size_t indexSpeciesPhase,
 					     vcs_VolPhase *owning) :
   IndexPhase(indexPhase),
   IndexSpeciesPhase(indexSpeciesPhase),
