+++ Cantera/src/thermo/IdealMolalSoln.h
@@ -459,7 +459,7 @@ namespace Cantera {
      *
      * @param k  Species index 
      */
-    virtual doublereal standardConcentration(int k=0) const;
+    virtual doublereal standardConcentration(size_t k=0) const;
 
     /*!
      * Returns the natural logarithm of the standard 
@@ -467,7 +467,7 @@ namespace Cantera {
      *
      * @param k  Species index 
      */
-    virtual doublereal logStandardConc(int k=0) const;
+    virtual doublereal logStandardConc(size_t k=0) const;
 
     /*!
      * Returns the units of the standard and generalized
