+++ Cantera/src/thermo/MolalityVPSSTP.h
@@ -477,7 +477,7 @@ namespace Cantera {
      *
      * @param k species index. Defaults to zero.
      */
-    virtual doublereal standardConcentration(int k=0) const;
+    virtual doublereal standardConcentration(size_t k=0) const;
 
     /**
      * Returns the natural logarithm of the standard 
@@ -485,7 +485,7 @@ namespace Cantera {
      *
      * @param k  species index
      */
-    virtual doublereal logStandardConc(int k=0) const;
+    virtual doublereal logStandardConc(size_t k=0) const;
 
     /**
      * Returns the units of the standard and generalized
@@ -873,7 +873,7 @@ namespace Cantera {
      *   law. This is the indentity of the Cl- species for the PHSCALE_NBS
      *   scaling
      */
-    int        m_indexCLM;
+    size_t m_indexCLM;
 
     //! Molecular weight of the Solvent
     doublereal m_weightSolvent;
